Add "no_value" and visual notification when the metric hasn't value
This commit adds a visual feature to the assessment dendrogram,
when a metric has not value, now it shows a "no_value" string and
a red color. Moreover, the parent will not take into account this
"no_value" metric and will show the string "( ! )" in order to
notice the user that the assessment is not complete and the
text will be in red color as well.

Signed-off-by: David Moreno <dmoreno@bitergia.com>